requirements.txt
python-dotenv
apache-beam
apache-beam[gcp]